Add new demo movie
