+++ README.md
@@ -12,7 +12,7 @@ npm install -g @compositor/x0
 [build-badge]: https://img.shields.io/travis/c8r/x0/master.svg?style=flat-square
 [build]: https://travis-ci.org/c8r/x0
 
-<img src='docs/demo.gif' class='demo-image' />
+<img src='docs/hello-x0.gif' class='demo-image' />
 
 ## Features
